proj-kaggle-fishes/src/features/VariationalAutoencoder.py
'''This script demonstrates how to build a variational autoencoder with Keras.

Reference: "Auto-Encoding Variational Bayes" https://arxiv.org/abs/1312.6114
'''
import numpy as np
import json
import matplotlib.pyplot as plt
from scipy.stats import norm
from skimage import io

from keras.layers import Input, Dense, Lambda
from keras.models import Model
from keras import backend as K
from keras import objectives
from keras.datasets import mnist
from keras.utils.np_utils import to_categorical

import os.path as op

from src.data import DataModel as dm

batch_size = 100
original_dim = 50176
latent_dim = 2
intermediate_dim = 250
nb_epoch = 50
epsilon_std = 1.0

x = Input(batch_shape=(None, original_dim))
h = Dense(intermediate_dim, activation='relu')(x)
z_mean = Dense(latent_dim)(h)
z_log_var = Dense(latent_dim)(h)


def sampling(args):
    z_mean, z_log_var = args
    epsilon = K.random_normal(shape=(batch_size, latent_dim), mean=0.,
                              std=epsilon_std)
    return z_mean + K.exp(z_log_var / 2) * epsilon

# note that "output_shape" isn't necessary with the TensorFlow backend
z = Lambda(sampling, output_shape=(latent_dim,))([z_mean, z_log_var])

# we instantiate these layers separately so as to reuse them later
decoder_h = Dense(intermediate_dim, activation='relu')
decoder_mean = Dense(original_dim, activation='sigmoid')
h_decoded = decoder_h(z)
x_decoded_mean = decoder_mean(h_decoded)


def vae_loss(x, x_decoded_mean):
    xent_loss = original_dim * objectives.binary_crossentropy(x, x_decoded_mean)
    kl_loss = - 0.5 * K.sum(1 + z_log_var - K.square(z_mean) - K.exp(z_log_var), axis=-1)
    return xent_loss + kl_loss

vae = Model(x, x_decoded_mean)
vae.compile(optimizer='rmsprop', loss=vae_loss)

# train the VAE on MNIST digits

f = dm.ProjFolder()
with open(op.join(f.data_processed, 'training_images.json'), 'rb') as file:
    training_img = json.load(file)


def split_data_rotated_cropped(training_img, f):
    X_train = []
    X_val = []
    y_train = []
    y_val = []
    for k, img in training_img.iteritems():
        im = io.imread(img['imgrotatecrop'], as_grey=True)
        if img['validation'] is False:
            while len(X_train) < 2700:
                X_train.append(im)
                y_train.append(img['fishtype'])
        else:
            while len(X_val) < 900:
                X_val.append(im)
                y_val.append(img['fishtype'])

    X_train = np.array(X_train)
    X_val = np.array(X_val)
    y_train = np.array(y_train)
    y_val = np.array(y_val)
    return(X_train, X_val, y_train, y_val)


split = split_data_rotated_cropped(training_img=training_img, f=f)
x_train, x_test, y_train_cat, y_test_cat = split[0], split[1], split[2], split[3]

x_train = x_train.astype('float32') / 255.
x_test = x_test.astype('float32') / 255.
x_train = x_train.reshape((len(x_train), np.prod(x_train.shape[1:])))
x_test = x_test.reshape((len(x_test), np.prod(x_test.shape[1:])))

y_train = []
y_test = []
classes = {'ALB':0,
           'BET':1,
                'DOL':2,
                'LAG':3,
                'NoF':4,
                'OTHER':5,
                'SHARK':6,
                'YFT':7}
for cat in y_train_cat:
    y_train.append(classes[cat])
for cat in y_test_cat:
    y_test.append(classes[cat])

y_train = to_categorical(y_train)

y_test = to_categorical(y_test)

vae.fit(x_train, x_train,
        shuffle=True,
        nb_epoch=nb_epoch,
        batch_size=batch_size,
        validation_data=(x_test, x_test))

# build a model to project inputs on the latent space
encoder = Model(x, z_mean)

# display a 2D plot of the digit classes in the latent space
x_test_encoded = encoder.predict(x_test, batch_size=batch_size)
plt.figure(figsize=(6, 6))
plt.scatter(x_test_encoded[:, 0], x_test_encoded[:, 1], c=np.array([x_test_encoded[:, 0], y_test]))
plt.colorbar()
plt.show()

# build a digit generator that can sample from the learned distribution
decoder_input = Input(shape=(latent_dim,))
_h_decoded = decoder_h(decoder_input)
_x_decoded_mean = decoder_mean(_h_decoded)
generator = Model(decoder_input, _x_decoded_mean)

# display a 2D manifold of the digits
n = 15  # figure with 15x15 digits
digit_size = 250
figure = np.zeros((digit_size * n, digit_size * n))
# linearly spaced coordinates on the unit square were transformed through the inverse CDF (ppf) of the Gaussian
# to produce values of the latent variables z, since the prior of the latent space is Gaussian
grid_x = norm.ppf(np.linspace(0.05, 0.95, n))
grid_y = norm.ppf(np.linspace(0.05, 0.95, n))

for i, yi in enumerate(grid_x):
    for j, xi in enumerate(grid_y):
        z_sample = np.array([[xi, yi]])
        x_decoded = generator.predict(z_sample)
        digit = x_decoded[0].reshape(digit_size, digit_size)
        figure[i * digit_size: (i + 1) * digit_size,
               j * digit_size: (j + 1) * digit_size] = digit

plt.figure(figsize=(10, 10))
plt.imshow(figure, cmap='Greys_r')
plt.show()
