proj-kaggle-fishes/requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
json
sys
os
glob
hashlib
numpy
pandas
cv2
time
pickle
collections
sklearn
time
skimage
mahotas
keras
math
cv2>=3.2.0
xgboost
re
tensorflow-gpu
sklearn
pandas
ipdb
opencv
h5py
